vocabulary_reminder/requirements.txt
pyTelegramBotAPI==3.6.6
pyyaml

# If this packafe fails: install sudo apt-get install unixodbc unixodbc-dev
#pyodbc
#python-sqlite-orm
